Create History component
