+++ client/src/observables/history.js
@@ -1,0 +1,7 @@
+import Vue from 'vue'
+
+export default Vue.observable({
+    // Used in HistoryModal.vue
+    showModal: false,
+    selectedHistoryRow: null,
+})
